refactor(web): share top picks and AI setup layouts
Replace duplicated top-picks pages and AI setup card grids with shared components while keeping route exports stable.
