+++ apps/web/src/pages/top-picks/TopPicksSeries.tsx
@@ -1,326 +1,5 @@
-import React, { useEffect, useState, useCallback } from 'react'
-import { useTranslation } from 'react-i18next'
-import { useNavigate } from 'react-router-dom'
-import {
-  Box,
-  Typography,
-  Grid,
-  Skeleton,
-  Alert,
-  Chip,
-  Card,
-  CardContent,
-  ToggleButton,
-  ToggleButtonGroup,
-} from '@mui/material'
-import TvIcon from '@mui/icons-material/Tv'
-import GridViewIcon from '@mui/icons-material/GridView'
-import ViewListIcon from '@mui/icons-material/ViewList'
-import PeopleIcon from '@mui/icons-material/People'
-import PlayArrowIcon from '@mui/icons-material/PlayArrow'
-import { MoviePoster, RankBadge } from '@aperture/ui'
-import { useUserRatings } from '../../hooks/useUserRatings'
-import { useWatching } from '../../hooks/useWatching'
-
-interface PopularSeries {
-  seriesId: string
-  title: string
-  year: number | null
-  posterUrl: string | null
-  backdropUrl: string | null
-  overview: string | null
-  genres: string[]
-  communityRating: number | null
-  network: string | null
-  uniqueViewers: number
-  totalEpisodesWatched: number
-  avgCompletionRate: number
-  popularityScore: number
-  rank: number
-}
-
-interface TopPicksConfig {
-  timeWindowDays: number
-  seriesCount: number
-  minUniqueViewers: number
-  lastRefreshedAt: string | null
-}
+import { TopPicksMediaPage } from './TopPicksMediaPage'
 
 export function TopPicksSeriesPage() {
-  const { t } = useTranslation()
-  const navigate = useNavigate()
-  const { getRating, setRating } = useUserRatings()
-  const { isWatching, toggleWatching } = useWatching()
-  const [series, setSeries] = useState<PopularSeries[]>([])
-  const [config, setConfig] = useState<TopPicksConfig | null>(null)
-  const [loading, setLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
-  const [viewMode, setViewMode] = useState<'grid' | 'list'>('grid')
-
-  const handleRate = useCallback(
-    async (seriesId: string, rating: number | null) => {
-      try {
-        await setRating('series', seriesId, rating)
-      } catch (err) {
-        console.error('Failed to rate series:', err)
-      }
-    },
-    [setRating]
-  )
-
-  useEffect(() => {
-    const fetchTopSeries = async () => {
-      try {
-        const response = await fetch('/api/top-picks/series', { credentials: 'include' })
-        if (response.ok) {
-          const data = await response.json()
-          setSeries(data.series)
-          setConfig(data.config)
-          setError(null)
-        } else {
-          setError(t('topPicksPage.errorLoadSeries'))
-        }
-      } catch {
-        setError(t('topPicksPage.errorConnect'))
-      } finally {
-        setLoading(false)
-      }
-    }
-
-    fetchTopSeries()
-  }, [t])
-
-  if (loading) {
-    return (
-      <Box>
-        <Skeleton variant="text" width={300} height={48} sx={{ mb: 1 }} />
-        <Skeleton variant="text" width={200} height={24} sx={{ mb: 4 }} />
-        <Grid container spacing={2}>
-          {[...Array(10)].map((_, i) => (
-            <Grid item xs={6} sm={4} md={3} lg={2} key={i}>
-              <Skeleton variant="rectangular" sx={{ width: '100%', aspectRatio: '2/3', borderRadius: 1 }} />
-            </Grid>
-          ))}
-        </Grid>
-      </Box>
-    )
-  }
-
-  return (
-    <Box>
-      {/* Header */}
-      <Box display="flex" justifyContent="space-between" alignItems="flex-start" mb={4} flexWrap="wrap" gap={2}>
-        <Box>
-          <Box display="flex" alignItems="center" gap={2} mb={1}>
-            <TvIcon sx={{ color: '#8b5cf6', fontSize: 32 }} />
-            <Typography variant="h4" fontWeight={700}>
-              {t('topPicksPage.pageTitleSeries')}
-            </Typography>
-          </Box>
-          <Typography variant="body1" color="text.secondary">
-            {t('topPicksPage.subtitleSeries')}
-            {config && t('topPicksPage.inLastDays', { count: config.timeWindowDays })}
-          </Typography>
-          {config?.lastRefreshedAt && (
-            <Typography variant="caption" color="text.secondary" display="block" mt={0.5}>
-              {t('topPicksPage.lastUpdated', {
-                date: new Date(config.lastRefreshedAt).toLocaleDateString(),
-              })}
-            </Typography>
-          )}
-        </Box>
-        <ToggleButtonGroup
-          value={viewMode}
-          exclusive
-          onChange={(_, v) => v && setViewMode(v)}
-          size="small"
-        >
-          <ToggleButton value="grid">
-            <GridViewIcon fontSize="small" />
-          </ToggleButton>
-          <ToggleButton value="list">
-            <ViewListIcon fontSize="small" />
-          </ToggleButton>
-        </ToggleButtonGroup>
-      </Box>
-
-      {error && (
-        <Alert severity="error" sx={{ mb: 3, borderRadius: 2 }}>
-          {error}
-        </Alert>
-      )}
-
-      {series.length === 0 && !error ? (
-        <Alert severity="info" sx={{ borderRadius: 2 }}>
-          {t('topPicksPage.emptySeries')}
-        </Alert>
-      ) : viewMode === 'grid' ? (
-        <Grid container spacing={2}>
-          {series.map((show) => (
-            <Grid item xs={6} sm={4} md={3} lg={2} key={show.seriesId}>
-              <Box position="relative">
-                <MoviePoster
-                  title={show.title}
-                  year={show.year}
-                  posterUrl={show.posterUrl}
-                  genres={show.genres}
-                  rating={show.communityRating}
-                  overview={show.overview}
-                  userRating={getRating('series', show.seriesId)}
-                  onRate={(rating) => handleRate(show.seriesId, rating)}
-                  responsive
-                  isWatching={isWatching(show.seriesId)}
-                  onWatchingToggle={() => toggleWatching(show.seriesId)}
-                  size="medium"
-                  onClick={() => navigate(`/series/${show.seriesId}`)}
-                />
-                <RankBadge rank={show.rank} size="large" />
-                {/* Viewers badge */}
-                <Chip
-                  icon={<PeopleIcon sx={{ fontSize: 14 }} />}
-                  label={show.uniqueViewers}
-                  size="small"
-                  sx={{
-                    position: 'absolute',
-                    bottom: 8,
-                    right: 8,
-                    backgroundColor: 'rgba(0,0,0,0.75)',
-                    color: 'white',
-                    fontSize: '0.7rem',
-                    height: 24,
-                    '& .MuiChip-icon': { color: 'white' },
-                  }}
-                />
-                {/* Network badge */}
-                {show.network && (
-                  <Chip
-                    label={show.network}
-                    size="small"
-                    sx={{
-                      position: 'absolute',
-                      top: 8,
-                      right: 8,
-                      backgroundColor: 'rgba(139, 92, 246, 0.9)',
-                      color: 'white',
-                      fontSize: '0.65rem',
-                      height: 20,
-                      maxWidth: 80,
-                      '& .MuiChip-label': { 
-                        overflow: 'hidden',
-                        textOverflow: 'ellipsis',
-                      },
-                    }}
-                  />
-                )}
-              </Box>
-            </Grid>
-          ))}
-        </Grid>
-      ) : (
-        <Box display="flex" flexDirection="column" gap={2}>
-          {series.map((show) => (
-            <Card
-              key={show.seriesId}
-              sx={{
-                backgroundColor: 'background.paper',
-                borderRadius: 2,
-                transition: 'transform 0.2s, box-shadow 0.2s',
-                '&:hover': {
-                  transform: 'translateY(-2px)',
-                  boxShadow: 4,
-                },
-              }}
-            >
-              <CardContent sx={{ display: 'flex', gap: 3, p: 2 }}>
-                {/* Rank */}
-                <RankBadge rank={show.rank} size="xlarge" absolute={false} />
-
-                {/* Poster */}
-                <Box
-                  component="img"
-                  src={show.posterUrl || undefined}
-                  alt={show.title}
-                  sx={{
-                    width: 80,
-                    height: 120,
-                    objectFit: 'cover',
-                    borderRadius: 1,
-                    backgroundColor: 'grey.800',
-                    flexShrink: 0,
-                  }}
-                />
-
-                {/* Info */}
-                <Box flex={1} minWidth={0}>
-                  <Box display="flex" alignItems="center" gap={1}>
-                    <Typography variant="h6" fontWeight={600} noWrap>
-                      {show.title}
-                    </Typography>
-                    {show.network && (
-                      <Chip 
-                        label={show.network} 
-                        size="small" 
-                        sx={{ 
-                          backgroundColor: 'rgba(139, 92, 246, 0.2)',
-                          fontSize: '0.7rem',
-                          height: 22,
-                        }} 
-                      />
-                    )}
-                  </Box>
-                  <Typography variant="body2" color="text.secondary" mb={1}>
-                    {show.year} {show.genres?.length > 0 && `• ${show.genres.slice(0, 3).join(', ')}`}
-                  </Typography>
-                  <Typography
-                    variant="body2"
-                    color="text.secondary"
-                    sx={{
-                      display: '-webkit-box',
-                      WebkitLineClamp: 2,
-                      WebkitBoxOrient: 'vertical',
-                      overflow: 'hidden',
-                    }}
-                  >
-                    {show.overview || t('common.noDescription')}
-                  </Typography>
-                </Box>
-
-                {/* Stats */}
-                <Box 
-                  display="flex" 
-                  flexDirection="column" 
-                  alignItems="flex-end" 
-                  justifyContent="center"
-                  gap={1}
-                  flexShrink={0}
-                >
-                  <Chip
-                    icon={<PeopleIcon sx={{ fontSize: 16 }} />}
-                    label={t('topPicksPage.viewersCount', { count: show.uniqueViewers })}
-                    size="small"
-                    sx={{ backgroundColor: 'rgba(139, 92, 246, 0.2)' }}
-                  />
-                  <Chip
-                    icon={<PlayArrowIcon sx={{ fontSize: 16 }} />}
-                    label={t('topPicksPage.episodesWatched', { count: show.totalEpisodesWatched })}
-                    size="small"
-                    variant="outlined"
-                  />
-                  {show.avgCompletionRate > 0 && (
-                    <Typography variant="caption" color="text.secondary">
-                      {t('topPicksPage.avgCompletion', {
-                        pct: Math.round(show.avgCompletionRate * 100),
-                      })}
-                    </Typography>
-                  )}
-                </Box>
-              </CardContent>
-            </Card>
-          ))}
-        </Box>
-      )}
-    </Box>
-  )
+  return <TopPicksMediaPage mediaType="series" />
 }
-
-
